// types/enums/lottoEnums.ts
// Lotto enums 'enum'

export enum LottoCategory {
  LottoResults = 1,
  LottoPlus1 = 2,
  LottoPlus2 = 3,
  EuroMillions = 4,
  EuroMillionsPlus = 5,
  DailyMillion9pm = 11,
  DailyMillionPlus9pm = 12,
  DailyMillion2pm = 15,
  DailyMillionPlus2pm = 16,
}

export enum WinningCategory {
  Jackpot_T1 = 1,
  FivePlusTwo_T2 = 2,
  FivePlusOne_T1 = 3,
  FivePlusOne_T2 = 4,
  Five_T1 = 5,
  FourPlusOne_T1 = 6,
  FourPlusTwo_T2 = 7,
  FourPlusOne_T2 = 8,
  Four_T1 = 9,
  ThreePlusOne_T1 = 10,
  ThreePlusTwo_T2 = 11,
  TwoPlusTwo_T2 = 12,
  ThreePlusOne_T2 = 13,
  Three_T1 = 14,
  OnePlusTwo_T2 = 15,
  TwoPlusOne_T2 = 16,
  Two_T2 = 17,
  TwoPlusOne_T1 = 18,
  TopPrize_T1 = 19,
}